pyproject.toml
[tool.poetry]
name = "crac-client-web"
version = "0.1.0"
description = "Client web for connecting to crac server via gRPC"
authors = ["Raniero Albanesi"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.10"
crac-protobuf = { git = "https://github.com/ara-astronomia/crac-protobuf", tag="0.1.21.0" }
grpcio-tools = "^1.43.0"
Flask = "^2.0.3"
opencv-python = "^4.5.5"
python-dotenv = "^0.20.0"
plotly = "^5.7.0"
kaleido = "0.2.1"
PySimpleGUI = "^4.60.4"


[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
